events-bot/requirements.txt
python-telegram-bot==13.13
environs==9.5.0
Jinja2==3.1.2